Day 17: ...and here's the rest of the convolution approach.
(Gratuitous changes because I realized having an upper bound would be
useful.  On the other hand some of the index arithmetic is prettier now.)

+++ day17/src/main.rs
@@ -4,7 +4,7 @@ mod pascal;
 
 use std::io::{stdin,BufRead};
 
-pub type Vol = u32;
+pub type Vol = u8;
 pub type Num = u64;
 
 fn main() {
